ryd garrage

+++ cardatabase/frontend_React/carapp/src/components/carlist.js
@@ -41,9 +41,15 @@ function CarList(){
     }, []);
     
     const onDelClick =(url) => {
-        fetch(url, {method: 'DELETE'})
-        .then(response => {fetchCars(); setOpen(true);})
-        .catch(err => console.error(err));
+        if (window.confirm("Are you sure to delete?")) {
+            fetch(url, {method: 'DELETE'})
+            .then(response => {if(response.ok){fetchCars(); setOpen(true);}
+            else {
+                alert('Something went wrsong!');
+                }
+            })
+            .catch(err => console.error(err))
+            }
     }
     const [open, setOpen] = useState(false);
